+++ Assets/Scripts/Shared/UnitBehaviour/Orders/CancelOrder.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class CancelOrder : Order {
 
-    public override bool GetInRange(RTSGameObject performingUnit, RTSGameObjectManager rtsGameObjectManager, float dt)
+    public override bool GetInRange(RTSGameObject performingUnit, RTSGameObjectManager rtsGameObjectManager, int dt)
     {
         return true;
     }
